fix: dev: Major refactor of code.
