+++ .github/workflows/ci.yml
@@ -28,11 +28,8 @@ jobs:
           pip freeze
           pip check
 
-      - name: Bandit
-        run: bandit -r .
-
-      - name: Test
-        run: make test
+      - name: Run Make
+        run: make
 
       - name: Publish Code Coverage
         uses: paambaati/codeclimate-action@v3.0.0
