+++ testinfra_bdd/then/__init__.py
@@ -1,0 +1,1 @@
+"""The then packages/fixtures for testinfra-bdd."""
