+++ testinfra_bdd/then/package.py
@@ -1,0 +1,42 @@
+"""Then system package fixtures for testinfra-bdd."""
+from pytest_bdd import (
+    then,
+    parsers
+)
+
+
+@then(parsers.parse('the package state is {expected_status}'))
+@then(parsers.parse('the package is {expected_status}'))
+def the_package_status_is(expected_status, testinfra_bdd_host):
+    """
+    Check the status of a package (installed/absent).
+
+    Parameters
+    ----------
+    expected_status : str
+        Can be absent, installed or present.
+    testinfra_bdd_host : testinfra_bdd.fixture.TestinfraBDD
+        The test fixture.
+
+    Raises
+    ------
+    AssertError
+        When the package is not in the expected state.
+    """
+    status_lookup = {
+        'absent': False,
+        'installed': True,
+        'present': True
+    }
+    expected_to_be_installed = status_lookup[expected_status]
+    pkg = testinfra_bdd_host.package
+    actual_status = pkg.is_installed
+
+    if expected_to_be_installed:
+        message = f'Expected {pkg.name} to be {expected_status} on {testinfra_bdd_host.hostname} but it is absent.'
+
+    if actual_status:
+        message = f'Expected {pkg.name} to be absent on {testinfra_bdd_host.hostname} '
+        message += 'but it is installed ({pkg.version}).'
+
+    assert actual_status == expected_to_be_installed, message
