+++ tests/step_defs/test_example.py
@@ -2,17 +2,17 @@
 import os
 import pytest
 
+import testinfra_bdd
+
 from pytest_bdd import given
 from pytest_bdd import scenarios
-
-import testinfra_bdd.plugins
 
 scenarios('../features/example.feature')
 
 
 # Ensure that the PyTest fixtures provided in testinfra-bdd are available to
 # your test suite.
-pytest_plugins = testinfra_bdd.plugins.add_plugins()
+pytest_plugins = testinfra_bdd.PYTEST_MODULES
 
 
 @given('on GitHub Actions we skip tests')
