+++ tests/step_defs/test_issue21.py
@@ -1,4 +1,5 @@
 """Fix Issue 21 feature tests."""
+import testinfra_bdd
 
 from pytest_bdd import scenarios
 
@@ -6,4 +7,4 @@ scenarios('../features/issue21.feature')
 
 # Ensure that the PyTest fixtures provided in testinfra-bdd are available to
 # your test suite.
-pytest_plugins = ['testinfra_bdd']
+pytest_plugins = testinfra_bdd.PYTEST_MODULES
